refactor health handling
