+++ api/pkg/controller/cluster/launching.go
@@ -7,7 +7,6 @@ import (
 	"github.com/golang/glog"
 	kubermaticv1 "github.com/kubermatic/kubermatic/api/pkg/crd/kubermatic/v1"
 	kuberneteshelper "github.com/kubermatic/kubermatic/api/pkg/kubernetes"
-	"github.com/kubermatic/kubermatic/api/pkg/provider/kubernetes"
 	"github.com/kubermatic/kubermatic/api/pkg/resources"
 
 	"k8s.io/api/core/v1"
@@ -19,40 +18,6 @@ import (
 	"k8s.io/client-go/util/cert"
 	certutil "k8s.io/client-go/util/cert"
 )
-
-func (cc *Controller) clusterHealth(c *kubermaticv1.Cluster) (bool, *kubermaticv1.ClusterHealth, error) {
-	ns := kubernetes.NamespaceName(c.Name)
-	health := kubermaticv1.ClusterHealth{
-		ClusterHealthStatus: kubermaticv1.ClusterHealthStatus{},
-	}
-
-	type depInfo struct {
-		healthy *bool
-	}
-
-	healthMapping := map[string]*depInfo{
-		resources.ApiserverDeploymenName:          {healthy: &health.Apiserver},
-		resources.ControllerManagerDeploymentName: {healthy: &health.Controller},
-		resources.SchedulerDeploymentName:         {healthy: &health.Scheduler},
-		resources.MachineControllerDeploymentName: {healthy: &health.MachineController},
-	}
-
-	for name := range healthMapping {
-		healthy, err := cc.healthyDeployment(ns, name)
-		if err != nil {
-			return false, nil, fmt.Errorf("failed to get dep health %q: %v", name, err)
-		}
-		*healthMapping[name].healthy = healthy
-	}
-
-	var err error
-	health.Etcd, err = cc.healthyStatefulSet(ns, resources.EtcdStatefulSetName)
-	if err != nil {
-		return false, nil, fmt.Errorf("failed to get etcd health: %v", err)
-	}
-
-	return health.AllHealthy(), &health, nil
-}
 
 // ensureClusterReachable checks if the cluster is reachable via its external name
 func (cc *Controller) ensureClusterReachable(c *kubermaticv1.Cluster) error {
